made meson stuff work on gcc/linux, replaced fmodf and fabsf with fmod and fabs for proper c89

+++ Tarp/Tarp.h
@@ -3095,9 +3095,9 @@ TARP_LOCAL void _tpGLStroke(_tpGLPath * _path,
             }
 
             offsetIntoPattern = _style->dashOffset;
-            if (fabsf(offsetIntoPattern) >= patternLen)
+            if (fabs(offsetIntoPattern) >= patternLen)
             {
-                offsetIntoPattern = fmodf(offsetIntoPattern, patternLen);
+                offsetIntoPattern = fmod(offsetIntoPattern, patternLen);
             }
 
             dashStartState.startDashLen = -offsetIntoPattern;
